refactor: split House tests into separate test classes for default House layout; Serialize method; and constructors, getters, setters
