+++ HideAndSeekTestProject/TestHouse.cs
@@ -15,12 +15,11 @@ namespace HideAndSeek
 {
     /// <summary>
     /// House tests for:
-    /// - property getters and setters
     /// - parameterized constructor
-    /// - Serialize method
-    /// - default House layout
+    /// - property getters and setters
+    /// - LocationsWithoutHidingPlaces and LocationsWithHidingPlaces setter methods
     /// 
-    /// These are integration tests using Location and LocationWithHidingPlace.
+    /// These are integration tests using Location and LocationWithHidingPlace
     /// </summary>
     [TestFixture]
     public class TestHouse
@@ -546,380 +545,5 @@ namespace HideAndSeek
             return new House("my house", "DefaultHouse", entry,
                 new List<Location>() { entry, hallway }, new List<LocationWithHidingPlace>() { bedroom, kitchen, office });
         }
-
-        [Test]
-        [Category("House Serialize Success")]
-        public void Test_House_SerializeMethod_DefaultHouse()
-        {
-            house = TestHouse_TestData.GetDefaultHouse();
-            Assert.That(house.Serialize(), Is.EqualTo(TestHouse_TestData.DefaultHouse_Serialized));
-        }
-
-        [Test]
-        [Category("House Serialize Success")]
-        public void Test_House_SerializeMethod_CustomHouse_WithLocationsWithoutHidingPlaces()
-        {
-            // ARRANGE
-            // Initialize variable to expected serialized House text
-            string expectedSerializedHouse =
-                #region expected serialized House
-                "{" +
-                    "\"Name\":\"dream house\"" + "," + 
-                    "\"HouseFileName\":\"DreamHouse\"" + "," + 
-                    "\"PlayerStartingPoint\":\"Kitchen\"" + "," + 
-                    "\"LocationsWithoutHidingPlaces\":" +
-                    "[" +
-                        "{" +
-                            "\"Name\":\"Kitchen\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"North\":\"Bedroom\"" + "," + 
-                                "\"East\":\"Pantry\"" + "," + 
-                                "\"West\":\"Office\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"Name\":\"Exercise Room\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"South\":\"Bedroom\"" +
-                            "}" +
-                        "}" +
-                    "]" + "," + 
-                    "\"LocationsWithHidingPlaces\":" +
-                    "[" +
-                        "{" +
-                            "\"HidingPlace\":\"under the bed\"" + "," + 
-                            "\"Name\":\"Bedroom\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"South\":\"Kitchen\"" + "," + 
-                                "\"North\":\"Exercise Room\"" + "," + 
-                                "\"East\":\"Closet\"" + "," + 
-                                "\"West\":\"Bathroom\"" +
-                            "}" +
-                        "}" + "," + 
-                        "{" +
-                            "\"HidingPlace\":\"in a box\"" + "," + 
-                            "\"Name\":\"Pantry\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"West\":\"Kitchen\"" +
-                            "}" +
-                        "}" + "," + 
-                        "{" +
-                            "\"HidingPlace\":\"under the desk\"" + "," + 
-                            "\"Name\":\"Office\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Kitchen\"" + "," + 
-                                "\"Northeast\":\"Bathroom\"" +
-                            "}" +
-                        "}" + "," + 
-                        "{" +
-                            "\"HidingPlace\":\"between the coats\"" + "," + 
-                            "\"Name\":\"Closet\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"West\":\"Bedroom\"" +
-                            "}" +
-                        "}" + "," + 
-                        "{" +
-                            "\"HidingPlace\":\"in the tub\"" + "," + 
-                            "\"Name\":\"Bathroom\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Bedroom\"" + "," + 
-                                "\"West\":\"Sensory Room\"" + "," + 
-                                "\"Southwest\":\"Office\"" +
-                            "}" +
-                        "}" + "," + 
-                        "{" +
-                            "\"HidingPlace\":\"under the bean bags\"" + "," + 
-                            "\"Name\":\"Sensory Room\"" + "," + 
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Bathroom\"" +
-                            "}" +
-                        "}" +
-                    "]" +
-                "}";
-            #endregion
-
-            #region create House
-            // Create starting point (Kitchen) and connect to new locations: Bedroom, Pantry, Office
-            Location kitchen = new Location("Kitchen");
-            LocationWithHidingPlace bedroom = kitchen.AddExit(Direction.North, "Bedroom", "under the bed");
-            LocationWithHidingPlace pantry = kitchen.AddExit(Direction.East, "Pantry", "in a box");
-            LocationWithHidingPlace office = kitchen.AddExit(Direction.West, "Office", "under the desk");
-
-            // Connect Bedroom to new locations: Exercise Room, Closet, Bathroom
-            Location exerciseRoom = bedroom.AddExit(Direction.North, "Exercise Room");
-            LocationWithHidingPlace closet = bedroom.AddExit(Direction.East, "Closet", "between the coats");
-            LocationWithHidingPlace bathroom = bedroom.AddExit(Direction.West, "Bathroom", "in the tub");
-
-            // Connect Office to new location: Sensory Room
-            LocationWithHidingPlace sensoryRoom = bathroom.AddExit(Direction.West, "Sensory Room", "under the bean bags");
-
-            // Connect Office to Bathroom
-            office.AddExit(Direction.Northeast, bathroom);
-
-            // Create enumerable of Location objects
-            IEnumerable<Location> locationsWithoutHidingPlaces = new List<Location>() { kitchen, exerciseRoom };
-
-            // Create enumerable of LocationWithHidingPlace objects
-            IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces = new List<LocationWithHidingPlace>()
-            {
-                bedroom, pantry, office, closet, bathroom, sensoryRoom
-            };
-
-            // Create House
-            house = new House("dream house", "DreamHouse", kitchen, locationsWithoutHidingPlaces, locationsWithHidingPlaces);
-            #endregion
-
-            // ACT
-            string serializedHouse = house.Serialize();
-
-            // Assert that serialized House text is as expected
-            Assert.That(serializedHouse, Is.EqualTo(expectedSerializedHouse));
-        }
-
-        [Test]
-        [Category("House Serialize Success")]
-        public void Test_House_SerializeMethod_CustomHouse_WithoutLocationsWithoutHidingPlaces()
-        {
-            // ARRANGE
-            // Initialize variable to expected serialized House text
-            string expectedSerializedHouse =
-            #region expected serialized House
-                "{" +
-                    "\"Name\":\"dream house\"" + "," +
-                    "\"HouseFileName\":\"DreamHouse\"" + "," +
-                    "\"PlayerStartingPoint\":\"Kitchen\"" + "," +
-                    "\"LocationsWithoutHidingPlaces\":[]" + "," +
-                    "\"LocationsWithHidingPlaces\":" +
-                    "[" +
-                        "{" +
-                            "\"HidingPlace\":\"in a cupboard\"" + "," +
-                            "\"Name\":\"Kitchen\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"North\":\"Bedroom\"" + "," +
-                                "\"East\":\"Pantry\"" + "," +
-                                "\"West\":\"Office\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"under the bed\"" + "," +
-                            "\"Name\":\"Bedroom\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"South\":\"Kitchen\"" + "," +
-                                "\"North\":\"Exercise Room\"" + "," +
-                                "\"East\":\"Closet\"" + "," +
-                                "\"West\":\"Bathroom\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"in a box\"" + "," +
-                            "\"Name\":\"Pantry\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"West\":\"Kitchen\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"under the desk\"" + "," +
-                            "\"Name\":\"Office\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Kitchen\"" + "," +
-                                "\"Northeast\":\"Bathroom\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"behind the balls\"" + "," +
-                            "\"Name\":\"Exercise Room\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"South\":\"Bedroom\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"between the coats\"" + "," +
-                            "\"Name\":\"Closet\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"West\":\"Bedroom\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"in the tub\"" + "," +
-                            "\"Name\":\"Bathroom\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Bedroom\"" + "," +
-                                "\"West\":\"Sensory Room\"" + "," +
-                                "\"Southwest\":\"Office\"" +
-                            "}" +
-                        "}" + "," +
-                        "{" +
-                            "\"HidingPlace\":\"under the bean bags\"" + "," +
-                            "\"Name\":\"Sensory Room\"" + "," +
-                            "\"ExitsForSerialization\":" +
-                            "{" +
-                                "\"East\":\"Bathroom\"" +
-                            "}" +
-                        "}" +
-                    "]" +
-                "}";
-            #endregion
-
-            #region create House
-            // Create starting point (Kitchen) and connect to new locations: Bedroom, Pantry, Office
-            LocationWithHidingPlace kitchen = new LocationWithHidingPlace("Kitchen", "in a cupboard");
-            LocationWithHidingPlace bedroom = kitchen.AddExit(Direction.North, "Bedroom", "under the bed");
-            LocationWithHidingPlace pantry = kitchen.AddExit(Direction.East, "Pantry", "in a box");
-            LocationWithHidingPlace office = kitchen.AddExit(Direction.West, "Office", "under the desk");
-
-            // Connect Bedroom to new locations: Exercise Room, Closet, Bathroom
-            LocationWithHidingPlace exerciseRoom = bedroom.AddExit(Direction.North, "Exercise Room", "behind the balls");
-            LocationWithHidingPlace closet = bedroom.AddExit(Direction.East, "Closet", "between the coats");
-            LocationWithHidingPlace bathroom = bedroom.AddExit(Direction.West, "Bathroom", "in the tub");
-
-            // Connect Office to new location: Sensory Room
-            LocationWithHidingPlace sensoryRoom = bathroom.AddExit(Direction.West, "Sensory Room", "under the bean bags");
-
-            // Connect Office to Bathroom
-            office.AddExit(Direction.Northeast, bathroom);
-
-            // Create enumerable of LocationWithHidingPlace objects
-            IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces = new List<LocationWithHidingPlace>()
-            {
-                kitchen, bedroom, pantry, office, exerciseRoom, closet, bathroom, sensoryRoom
-            };
-
-            // Create House
-            house = new House("dream house", "DreamHouse", kitchen, new List<Location>(), locationsWithHidingPlaces);
-            #endregion
-
-            // ACT
-            string serializedHouse = house.Serialize();
-
-            // Assert that serialized House text is as expected
-            Assert.That(serializedHouse, Is.EqualTo(expectedSerializedHouse));
-        }
-
-        #region Test default House layout
-        /// <summary>
-        /// Assert that layout of House is as expected using House's GetExit method and Location's Name property
-        /// 
-        /// CREDIT: adapted from HideAndSeek project's TestHouse class's TestLayout() test method
-        ///         © 2023 Andrew Stellman and Jennifer Greene
-        ///         Published under the MIT License
-        ///         https://github.com/head-first-csharp/fourth-edition/blob/master/Code/Chapter_10/HideAndSeek_part_3/HideAndSeekTests/TestHouse.cs
-        ///         Link valid as of 02-26-2025
-        ///         
-        /// CHANGES:
-        /// -I changed the method name to be consistent with the conventions I'm using in this test project.
-        /// -I put all the assertions in the body of a multiple assert so all assertions will be run.
-        /// -I changed the assertions to use the constraint model to stay up-to-date.
-        /// -I added some comments for easier reading.
-        /// -I added messages to the assertions to make them easier to debug.
-        /// </summary>
-        [Test]
-        [Category("House Layout GetExit Name")]
-        public void Test_House_Layout()
-        {
-            // Set House
-            house = TestHouse_TestData.GetDefaultHouse();
-
-            Assert.Multiple(() =>
-            {
-                // Assert that name of StartingPoint is correct
-                Assert.That(house.StartingPoint.Name, Is.EqualTo("Entry"), "name of starting point");
-
-                // Assert that Garage is located "Out" from StartingPoint
-                var garage = house.StartingPoint.GetExit(Direction.Out);
-                Assert.That(garage.Name, Is.EqualTo("Garage"), "Garage is Out from Entry");
-
-                // Assert that Hallway is located "East" of StartingPoint
-                var hallway = house.StartingPoint.GetExit(Direction.East);
-                Assert.That(hallway.Name, Is.EqualTo("Hallway"), "Hallway is East of Entry");
-
-                // Assert that Kitchen is located "Northwest" of Hallway
-                var kitchen = hallway.GetExit(Direction.Northwest);
-                Assert.That(kitchen.Name, Is.EqualTo("Kitchen"), "Kitchen is Northwest of Hallway");
-
-                // Assert that Bathroom is located "North" of Hallway
-                var bathroom = hallway.GetExit(Direction.North);
-                Assert.That(bathroom.Name, Is.EqualTo("Bathroom"), "Bathroom is North of Hallway");
-
-                // Assert that Living Room is located "South" of Hallway
-                var livingRoom = hallway.GetExit(Direction.South);
-                Assert.That(livingRoom.Name, Is.EqualTo("Living Room"), "Living Room is South of Hallway");
-
-                // Assert that Landing is located "Up" from Hallway
-                var landing = hallway.GetExit(Direction.Up);
-                Assert.That(landing.Name, Is.EqualTo("Landing"), "Landing is Up from Hallway");
-
-                // Assert that Master Bedroom is located "Northwest" of Landing
-                var masterBedroom = landing.GetExit(Direction.Northwest);
-                Assert.That(masterBedroom.Name, Is.EqualTo("Master Bedroom"), "Master Bedroom is Northwest of Landing");
-
-                // Assert that Master Bath is located "East" of Master Bedroom
-                var masterBath = masterBedroom.GetExit(Direction.East);
-                Assert.That(masterBath.Name, Is.EqualTo("Master Bath"), "Master Bath is East of Master Bedroom");
-
-                // Assert that Second Bathroom is located "West" of Landing
-                var secondBathroom = landing.GetExit(Direction.West);
-                Assert.That(secondBathroom.Name, Is.EqualTo("Second Bathroom"), "Second Bathroom is West of Landing");
-
-                // Assert that Nursery is located Southwest of Landing
-                var nursery = landing.GetExit(Direction.Southwest);
-                Assert.That(nursery.Name, Is.EqualTo("Nursery"), "Nursery is Southwest of Landing");
-
-                // Assert that Pantry is located South of Landing
-                var pantry = landing.GetExit(Direction.South);
-                Assert.That(pantry.Name, Is.EqualTo("Pantry"), "Pantry is South of Landing");
-
-                // Assert that Kids Room is located "Southeast" of Landing
-                var kidsRoom = landing.GetExit(Direction.Southeast);
-                Assert.That(kidsRoom.Name, Is.EqualTo("Kids Room"), "Kids Room is Southeast of Landing");
-
-                // Assert that Attic is located "Up" from Landing
-                var attic = landing.GetExit(Direction.Up);
-                Assert.That(attic.Name, Is.EqualTo("Attic"), "Attic is Up from Landing");
-            });
-        }
-
-        [TestCase("Entry")]
-        [TestCase("Hallway")]
-        [TestCase("Landing")]
-        [Category("House Locations LocationType")]
-        public void Test_House_Locations_AreNotOfType_LocationWithHidingPlace(string locationName)
-        {
-            house = TestHouse_TestData.GetDefaultHouse();
-            Assert.That(house.Locations.Where((l) => l.Name == locationName).First(), Is.Not.InstanceOf<LocationWithHidingPlace>());
-        }
-
-        [TestCase("Garage")]
-        [TestCase("Kitchen")]
-        [TestCase("Living Room")]
-        [TestCase("Bathroom")]
-        [TestCase("Master Bedroom")]
-        [TestCase("Master Bath")]
-        [TestCase("Second Bathroom")]
-        [TestCase("Kids Room")]
-        [TestCase("Nursery")]
-        [TestCase("Pantry")]
-        [TestCase("Attic")]
-        [Category("House Locations LocationWithHidingPlaceType")]
-        public void Test_House_Locations_AreOfType_LocationWithHidingPlace(string locationName)
-        {
-            house = TestHouse_TestData.GetDefaultHouse();
-            Assert.That(house.Locations.Where((l) => l.Name == locationName).First(), Is.InstanceOf<LocationWithHidingPlace>());
-        }
-        #endregion
     }
 }
