--- HideAndSeekTestProject/TestHouse_TestData.cs
@@ -1,212 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.IO.Abstractions;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HideAndSeek
-{
-    /// <summary>
-    /// Test data for some House tests
-    /// </summary>
-    public static class TestHouse_TestData
-    {
-        /// <summary>
-        /// Text representing default House for tests serialized
-        /// </summary>
-        public static string DefaultHouse_Serialized
-        {
-            get
-            {
-                return
-                    "{" +
-                        "\"Name\":\"my house\"" + "," +
-                        "\"HouseFileName\":\"DefaultHouse\"" + "," +
-                        "\"PlayerStartingPoint\":\"Entry\"" + "," +
-                        "\"LocationsWithoutHidingPlaces\":" +
-                        "[" +
-                            "{" +
-                                "\"Name\":\"Hallway\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"West\":\"Entry\"," +
-                                    "\"Northwest\":\"Kitchen\"," +
-                                    "\"North\":\"Bathroom\"," +
-                                    "\"South\":\"Living Room\"," +
-                                    "\"Up\":\"Landing\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"Name\":\"Landing\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Down\":\"Hallway\"," +
-                                    "\"Up\":\"Attic\"," +
-                                    "\"Southeast\":\"Kids Room\"," +
-                                    "\"Northwest\":\"Master Bedroom\"," +
-                                    "\"Southwest\":\"Nursery\"," +
-                                    "\"South\":\"Pantry\"," +
-                                    "\"West\":\"Second Bathroom\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"Name\":\"Entry\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Out\":\"Garage\"," +
-                                    "\"East\":\"Hallway\"" +
-                                "}" +
-                            "}" +
-                        "]" + "," +
-                        "\"LocationsWithHidingPlaces\":" +
-                        "[" +
-                            "{" +
-                                "\"HidingPlace\":\"in a trunk\"," +
-                                "\"Name\":\"Attic\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Down\":\"Landing\"" +
-                                "}" +
-                            "}," +
-                            "{\"HidingPlace\":\"behind the door\"," +
-                                "\"Name\":\"Bathroom\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"South\":\"Hallway\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"in the bunk beds\"," +
-                                "\"Name\":\"Kids Room\"," +
-                                "\"ExitsForSerialization\":" +
-                                    "{" +
-                                        "\"Northwest\":\"Landing\"" +
-                                    "}" +
-                                "}," +
-                            "{" +
-                                "\"HidingPlace\":\"under the bed\"," +
-                                "\"Name\":\"Master Bedroom\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Southeast\":\"Landing\"," +
-                                    "\"East\":\"Master Bath\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"behind the changing table\"," +
-                                "\"Name\":\"Nursery\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Northeast\":\"Landing\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"inside a cabinet\"," +
-                                "\"Name\":\"Pantry\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"North\":\"Landing\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"in the shower\"," +
-                                "\"Name\":\"Second Bathroom\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"East\":\"Landing\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"next to the stove\"," +
-                                "\"Name\":\"Kitchen\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"Southeast\":\"Hallway\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"in the tub\"," +
-                                "\"Name\":\"Master Bath\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"West\":\"Master Bedroom\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"behind the car\"," +
-                                "\"Name\":\"Garage\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"In\":\"Entry\"" +
-                                "}" +
-                            "}," +
-                            "{" +
-                                "\"HidingPlace\":\"behind the sofa\"," +
-                                "\"Name\":\"Living Room\"," +
-                                "\"ExitsForSerialization\":" +
-                                "{" +
-                                    "\"North\":\"Hallway\"" +
-                                "}" +
-                            "}" +
-                        "]" +
-                    "}";
-            }
-        }
-
-        /// <summary>
-        /// Get new House object for testing purposes
-        /// </summary>
-        /// <returns>House object for testing purposes</returns>
-        public static House GetDefaultHouse()
-        {
-            // Create Entry and connect to new locations: Garage, Hallway
-            Location entry = new Location("Entry");
-            LocationWithHidingPlace garage = entry.AddExit(Direction.Out, "Garage", "behind the car");
-            Location hallway = entry.AddExit(Direction.East, "Hallway");
-
-            // Connect Hallway to new locations: Kitchen, Bathroom, Living Room, Landing
-            LocationWithHidingPlace kitchen = hallway.AddExit(Direction.Northwest, "Kitchen", "next to the stove");
-            LocationWithHidingPlace bathroom = hallway.AddExit(Direction.North, "Bathroom", "behind the door");
-            LocationWithHidingPlace livingRoom = hallway.AddExit(Direction.South, "Living Room", "behind the sofa");
-            Location landing = hallway.AddExit(Direction.Up, "Landing");
-
-            // Connect Landing to new locations: Attic, Kids Room, Master Bedroom, Nursery, Pantry, Second Bathroom
-            LocationWithHidingPlace attic = landing.AddExit(Direction.Up, "Attic", "in a trunk");
-            LocationWithHidingPlace kidsRoom = landing.AddExit(Direction.Southeast, "Kids Room", "in the bunk beds");
-            LocationWithHidingPlace masterBedroom = landing.AddExit(Direction.Northwest, "Master Bedroom", "under the bed");
-            LocationWithHidingPlace nursery = landing.AddExit(Direction.Southwest, "Nursery", "behind the changing table");
-            LocationWithHidingPlace pantry = landing.AddExit(Direction.South, "Pantry", "inside a cabinet");
-            LocationWithHidingPlace secondBathroom = landing.AddExit(Direction.West, "Second Bathroom", "in the shower");
-
-            // Connect Master Bedroom to new location: Master Bath
-            LocationWithHidingPlace masterBath = masterBedroom.AddExit(Direction.East, "Master Bath", "in the tub");
-
-            // Create list of Location objects (no hiding places)
-            IEnumerable<Location> locationsWithoutHidingPlaces = new List<Location>()
-            {
-                hallway, landing, entry
-            };
-
-            // Create list of LocationWithHidingPlace objects
-            IEnumerable<LocationWithHidingPlace> locationsWithHidingPlaces = new List<LocationWithHidingPlace>()
-            {
-                attic,
-                bathroom,
-                kidsRoom,
-                masterBedroom,
-                nursery,
-                pantry,
-                secondBathroom,
-                kitchen,
-                masterBath,
-                garage,
-                livingRoom
-            };
-
-            // Create and return new House
-            return new House("my house", "DefaultHouse", entry, locationsWithoutHidingPlaces, locationsWithHidingPlaces);
-        }
-    }
-}
